# Serialize gasnet polling calls for UCX too